Coupling diversification:
The forms folder is reintroduced from the original MASOM (requires a forms folder when training, even if empty). New to Spire Muse is that 20% of the times coupling mode is activated, the response will not be appropriate, but "rude". The rudeness consists in picking a random file from the forms folder as the FO training, and a random state in that FO as the starting point. The playback will respond with the best matching units available in the SOM.
Made the coupling mode more "sticky", minimum activation time 15 seconds.