Added write history to file
After End session is pressed, a CSV file will be produced in the Run folder. This file contains the entire session history. Rename the file immediately, so that the next session will not overwrite the existing file (todo: another naming scheme, so new file names will be numbered incrementally).

Todo #2: Loudness parameter needs to be multiplied by 20 for it to be correct in history (in runtime, the multiplication is done in SM_model.maxpat, but session history reads from the multislider in SM_view.maxpat).